requirements.txt
attrs==19.3.0
autopep8==1.5
awscli==1.18.45
awsebcli==3.17.1
bcrypt==3.1.7
blessed==1.17.2
boto3==1.12.17
botocore==1.15.45
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1
colorama==0.3.9
cryptography==2.8
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
dominate==2.5.1
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.16.0
idna==2.7
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
jsonschema==3.2.0
MarkupSafe==1.1.1
numpy==1.18.3
pandas==1.0.3
paramiko==2.7.1
pathspec==0.5.9
plotly==4.7.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.20
PyNaCl==1.3.0
pyrsistent==0.15.7
python-dateutil==2.8.0
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.2
requests==2.20.1
rsa==3.4.2
s3transfer==0.3.3
selenium==3.141.0
semantic-version==2.5.0
simplejson==3.17.0
six==1.11.0
SQLAlchemy==1.3.15
termcolor==1.1.0
texttable==1.6.2
urllib3==1.24.3
visitor==0.1.3
wcwidth==0.1.8
websocket-client==0.57.0
Werkzeug==1.0.0
wget==3.2
WTForms==2.2.1
xlrd==1.2.0
yamllint==1.20.0
yfinance==0.1.54
zipp==3.1.0
